edit lesson6
