+++ .gitignore
@@ -2,3 +2,4 @@
 *.iml
 /target
 *.mvn
+target/
