+++ lesson6/spring-shell.log
@@ -52,3 +52,10 @@
 1570376966659:faB
 1570376989647:aB
 1570377034821:faB
+1570506749605:l
+1570506755018:fidB
+1570506832588:stacktrace
+1570507154360:l
+1570507158171:fidB
+1570507343359:l
+1570507346097:faB
