+++ lesson6/src/main/java/amandzhi/springjdbc/SpringjdbcApplication.java
@@ -1,6 +1,6 @@
 package amandzhi.springjdbc;
 
-import amandzhi.springjdbc.repository.AuthorRepositoryDaoImpl;
+import amandzhi.springjdbc.repository.AuthorRepositoryImpl;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -10,7 +10,7 @@ public class SpringjdbcApplication {
 
     public static void main(String[] args) throws InterruptedException {
         ConfigurableApplicationContext context = SpringApplication.run(SpringjdbcApplication.class, args);
-        AuthorRepositoryDaoImpl bean = context.getBean(AuthorRepositoryDaoImpl.class);
+        AuthorRepositoryImpl bean = context.getBean(AuthorRepositoryImpl.class);
         System.out.println(bean.findAll().size());
     }
 }
