+++ lesson6/src/main/java/amandzhi/springjdbc/command/Commands.java
@@ -5,7 +5,9 @@ import amandzhi.springjdbc.io.IOService;
 import amandzhi.springjdbc.model.Author;
 import amandzhi.springjdbc.model.Book;
 import amandzhi.springjdbc.model.Genre;
-import amandzhi.springjdbc.service.Service;
+import amandzhi.springjdbc.service.AuthorService;
+import amandzhi.springjdbc.service.BookService;
+import amandzhi.springjdbc.service.GenreService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.MessageSource;
@@ -20,14 +22,15 @@ import java.util.Collection;
 @ShellComponent
 public class Commands {
     private String userName;
-    private Service<Genre> genreService;
-    private Service<Author> authorService;
-    private Service<Book> bookService;
+    private GenreService<Genre> genreService;
+    private AuthorService<Author> authorService;
+    private BookService<Book> bookService;
     private IOService iOService;
     private LocaleProps localeProps;
     private MessageSource messageSource;
     private final String WELCOME = "welcome";
     private final String QUESTION_NAME_BOOK = "questionNameBook";
+    private final String QUESTION_ID_BOOK = "questionIdBook";
     private final String QUESTION_NAME_AUTHOR = "questionAuthorBook";
     private final String QUESTION_NAME_GENRE = "questionGenreBook";
     private final String LOGIN_ENTER = "enterLogin";
@@ -38,7 +41,7 @@ public class Commands {
     private final String ENTER_ID_BOOK = "enterIdBook";
 
     @Autowired
-    public Commands(@Qualifier("genreService") Service<Genre> genreService, @Qualifier("authorService") Service<Author> authorService,@Qualifier("bookService") Service<Book> bookService, IOService iOService, LocaleProps localeProps, MessageSource messageSource) {
+    public Commands(GenreService<Genre> genreService, AuthorService<Author> authorService, BookService<Book> bookService, IOService iOService, LocaleProps localeProps, MessageSource messageSource) {
         this.genreService = genreService;
         this.authorService = authorService;
         this.bookService = bookService;
@@ -69,13 +72,9 @@ public class Commands {
         String genreName = iOService.readString();
         Book book = new Book();
         book.setTitle(bookTitle);
-        Author author = new Author();
-        author.setName(authorName);
-        Genre genre = new Genre();
-        genre.setName(genreName);
-        Author authorFind = authorService.findByName(author.getName());
+        Author authorFind = authorService.findByName(authorName);
         book.setAuthor(authorFind);
-        Genre genreFind = genreService.findByName(genre.getName());
+        Genre genreFind = genreService.findByName(genreName);
         book.setGenre(genreFind);
         bookService.insert(book);
         iOService.printString(messageSource.getMessage(SUCCESS, null, localeProps.getLocale()));
@@ -87,6 +86,15 @@ public class Commands {
         iOService.printString(messageSource.getMessage(QUESTION_NAME_BOOK, null, localeProps.getLocale()));
         String bookTitle = iOService.readString();
         Book bookFind = bookService.findByName(bookTitle);
+        iOService.printString(bookFind.getTitle());
+    }
+
+    @ShellMethod(value = "find by id Book", key = {"findByIdBook", "fidB"})
+    @ShellMethodAvailability(value = "isLogin")
+    public void findByIdBook() {
+        iOService.printString(messageSource.getMessage(QUESTION_ID_BOOK, null, localeProps.getLocale()));
+        String bookId = iOService.readString();
+        Book bookFind = bookService.findById(Long.parseLong(bookId));
         iOService.printString(bookFind.getTitle());
     }
