+++ lesson6/src/main/java/amandzhi/springjdbc/model/Author.java
@@ -1,9 +1,17 @@
 package amandzhi.springjdbc.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class Author {
     private Long id;
     private String name;
+
+    public Author(String name) {
+        this.name = name;
+    }
 }
