+++ lesson6/src/main/java/amandzhi/springjdbc/model/Book.java
@@ -1,8 +1,12 @@
 package amandzhi.springjdbc.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@AllArgsConstructor
+@NoArgsConstructor
 public class Book {
     private Long id;
     private String title;
