+++ lesson6/src/main/java/amandzhi/springjdbc/model/Genre.java
@@ -1,9 +1,17 @@
 package amandzhi.springjdbc.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class Genre {
     private Long id;
     private String name;
+
+    public Genre(String name) {
+        this.name = name;
+    }
 }
