+++ lesson6/src/main/java/amandzhi/springjdbc/repository/GenreRepository.java
@@ -1,0 +1,12 @@
+package amandzhi.springjdbc.repository;
+
+import java.util.Collection;
+
+public interface GenreRepository<Genre> {
+    boolean insert(Genre entity);
+    boolean update(Genre entity);
+    Genre findById(Long id);
+    boolean deleteById(Long id);
+    Genre findByName(String name);
+    Collection<Genre> findAll();
+}
