+++ lesson6/src/main/java/amandzhi/springjdbc/service/AuthorService.java
@@ -1,0 +1,12 @@
+package amandzhi.springjdbc.service;
+
+import java.util.Collection;
+
+public interface AuthorService<Author> {
+    boolean insert(Author entity);
+    boolean update(Author entity);
+    Author findById(Long id);
+    boolean deleteById(Long id);
+    Author findByName(String name);
+    Collection<Author> findAll();
+}
