+++ lesson6/src/main/resources/i18n/bundle_en.properties
@@ -1,6 +1,7 @@
 success=Success
 welcome=welcome
 questionNameBook=Please enter book name
+questionIdBook=Please enter book id
 questionAuthorBook=Please enter book author
 questionGenreBook=Please enter book genre
 enterLogin=Please enter login command
