+++ lesson6/src/main/resources/i18n/bundle_ru_RU.properties
@@ -1,6 +1,7 @@
 success= Успешно
 welcome=Добро пожаловать
 questionNameBook=Введите название книги
+questionIdBook=Введите id книги
 questionAuthorBook=Введите имя автора книги
 questionGenreBook=Введите название жанра книги
 enterLogin=Залогиньтесь
