+++ lesson6/src/test/java/amandzhi/springjdbc/repository/BookRepositoryImplTest.java
@@ -1,0 +1,106 @@
+package amandzhi.springjdbc.repository;
+
+import amandzhi.springjdbc.model.Author;
+import amandzhi.springjdbc.model.Book;
+import amandzhi.springjdbc.model.Genre;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.util.Collection;
+
+@DisplayName("DAO для книг")
+@ExtendWith(SpringExtension.class)
+@JdbcTest
+@Import(value = {BookRepositoryImpl.class, AuthorRepositoryImpl.class, GenreRepositoryImpl.class})
+class BookRepositoryImplTest {
+
+    @Autowired
+    private BookRepositoryImpl bookRepositoryDao;
+
+    @Autowired
+    private AuthorRepository<Author> authorRepository;
+
+    @Autowired
+    private GenreRepository<Genre> genreRepository;
+
+    @Test
+    void insert() {
+        Author author = new Author();
+        author.setName("testjamandzhi");
+        author.setId(11L);
+        Genre genre = new Genre();
+        genre.setName("testfantasy");
+        genre.setId(11L);
+        Book book = new Book();
+        book.setTitle("newMyBook");
+        book.setAuthor(author);
+        book.setGenre(genre);
+        boolean insert = bookRepositoryDao.insert(book);
+        Book findBook = bookRepositoryDao.findByName(book.getTitle());
+        Assertions.assertTrue(insert);
+        Assertions.assertNotNull(findBook);
+    }
+
+    @Test
+    void update() {
+        Author author = new Author();
+        author.setName("testjamandzhi");
+        author.setId(11L);
+        Genre genre = new Genre();
+        genre.setName("testfantasy");
+        genre.setId(11L);
+        Book book = new Book();
+        book.setId(24L);
+        book.setTitle("newUpdateMyBook");
+        book.setAuthor(author);
+        book.setGenre(genre);
+        boolean update = bookRepositoryDao.update(book);
+        Book findBook = bookRepositoryDao.findByName(book.getTitle());
+        Assertions.assertTrue(update);
+        Assertions.assertNotNull(findBook);
+    }
+
+
+    @Test
+    void findById() {
+        Book book = bookRepositoryDao.findById(22L);
+        Assertions.assertNotNull(book);
+        Assertions.assertNotNull(book.getTitle());
+        Assertions.assertNotNull(book.getAuthor());
+        Assertions.assertNotNull(book.getGenre());
+    }
+
+    @Test
+    void deleteById() {
+        boolean delete = bookRepositoryDao.deleteById(23L);
+        Assertions.assertTrue(delete);
+    }
+
+    @Test
+    void findByName() {
+        Book book = bookRepositoryDao.findByName("testjamandzhi");
+        Assertions.assertNotNull(book);
+        Assertions.assertNotNull(book.getTitle());
+        Assertions.assertNotNull(book.getAuthor());
+        Assertions.assertNotNull(book.getGenre());
+    }
+
+    @Test
+    void findAll() {
+        Collection<Book> all = bookRepositoryDao.findAll();
+        Assertions.assertEquals(3, all.size());
+        for (Book b: all) {
+            Assertions.assertNotNull(b);
+            Assertions.assertNotNull(b.getTitle());
+            Assertions.assertNotNull(b.getAuthor());
+            Assertions.assertNotNull(b.getGenre());
+        }
+    }
+}
